refactor: rename benchmark_base.py to base.py
