+++ src/confopt/benchmark/__init__.py
@@ -1,4 +1,4 @@
-from .benchmark_base import BenchmarkBase
+from .base import BenchmarkBase
 from .nb101_benchmark import NB101Benchmark
 from .nb201_benchmark import NB201Benchmark
 from .nb301_benchmark import NB301Benchmark
